Bloqueando exclusão de especialidade que esteja presente em alguma consulta agendada

+++ src/main/java/com/mballem/curso/security/web/controller/MedicoController.java
@@ -66,8 +66,14 @@ public class MedicoController {
     @GetMapping({"/id/{idMed}/excluir/especializacao/{idEsp}"})
     public String excluirEspecialidadePorMedico(@PathVariable("idMed") Long idMed,
     					 @PathVariable("idEsp") Long idEsp, RedirectAttributes attr){
-        service.excluirEspecialidadePorMedico(idMed, idEsp);
-        attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso!");
+
+        if (service.existeEspecialidadeAgendada(idMed, idEsp)){
+            attr.addFlashAttribute("falha", "Não foi possivel realizar a exclusão, tem consultas agendadas com essa especialidade!");
+        }else {
+
+            service.excluirEspecialidadePorMedico(idMed, idEsp);
+            attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso!");
+        }
         return "redirect:/medicos/dados";
     }
